refactor: componentizando melhor o codigo
