+++ src/components/Main/ModeToggle/index.tsx
@@ -1,0 +1,32 @@
+import { Button } from "@/components/ui/button";
+import { Moon, Sun } from "lucide-react";
+import { useTheme } from "next-themes";
+import * as React from "react"
+
+export const ModeToggle = () => {
+    const { setTheme, resolvedTheme } = useTheme()
+    const [mounted, setMounted] = React.useState(false)
+
+    React.useEffect(() => {
+        setMounted(true)
+    }, [])
+
+    if (!mounted) return null
+
+    const isDark = resolvedTheme === "dark"
+
+    const toggleTheme = () => {
+        setTheme(isDark ? "light" : "dark")
+    }
+
+    return (
+        <Button variant="ghost" size="icon" onClick={toggleTheme} className="cursor-pointer">
+            {isDark ? (
+                <Sun className="h-[1.2rem] w-[1.2rem] transition-all" />
+            ) : (
+                <Moon className="h-[1.2rem] w-[1.2rem] transition-all" />
+            )}
+            <span className="sr-only">Toggle theme</span>
+        </Button>
+    )
+}
